flex optimizations, infura dedicated gateway

+++ events/messageCreate.js
@@ -19,19 +19,23 @@ module.exports = {
     let messageContent = message.content.toLowerCase();
 
     // Linguini L take
+    /*
     if (message.author.id === '932913898473013268') {
       Tenor.Search.Query('L', "50").then(results => {
         const randomGif = results[Math.floor(Math.random() * results.length)];
-        message.channel.send(randomGif.url);    
+        message.reply(randomGif.url);    
       })
     }
+    */
 
     // Linguini_ Kong Fud Filter
-    /*if (message.author.id === '932913898473013268' && (messageContent.includes('kong') || messageContent.includes('kongs') || messageContent.includes('ongs') || messageContent.includes('ngs'))) {
+    /*
+    if (message.author.id === '932913898473013268' && (messageContent.includes('kong') || messageContent.includes('kongs') || messageContent.includes('ongs') || messageContent.includes('ngs'))) {
       message.delete();
       message.channel.send('no no no - bad boy.');
       console.log(`Command: Delete Linguini_ Fud`)
-    }*/
+    }
+    */
 
     // Twitter Function
     // This will retrieve messages that include a tweet link, add fun reacts, and repost it in a separate channel
@@ -66,12 +70,7 @@ module.exports = {
       for (let searchQuery in matchMessage) {
         // Quick fixes for certain terms
         switch (matchMessage[searchQuery]) {
-          case 'bull':
-          case 'bear':
-            matchMessage[searchQuery] = 'crypto';
-            break;
           case 'ginger':
-          case 'gingers':
             matchMessage[searchQuery] = 'rip';
             break;
           case 'ups':
